requirements.txt
Django>=3.0
psycopg2-binary>=2.8
requests>=2.25
celery>=5.0
redis>=3.5.3
djangorestframework>=3.12
markdown>=3.3.3
djangorestframework-simplejwt==4.7.1
PyJWT==2.1.0